Add working project of an SVF filter plugin
Add working VST/VST3 plugin containing an 8-mode State Variable Filter
with GUI controls.

Note:
The GUI controls are basically copied from another project. So the Shelf
Gain knob doesn't have the correct values that the ShelvingFilter needs
right now. Also, the Resonance parameter needs to be shaped (internally)
so it has more control in the upper range (above 0.5 or so). There are
also other changes that could be made, but these are the most important
for now.