Create function to process sample block and ....
Create a function called processAudioBlock() that processes a sample
block, instead of single samples. Also change the name of
processChannel() to a more accurate name, processAudioSample().

At the moment, processAudioSample() returns a value, while
processAudioBlock() processes the buffer through a pointer argument.
Efforts should be made to make these two functions more similar in this
regard.

Add a function setIsActive() to set whether the filter is performing any
processing and setting parameters or not. If it is set to false, the
audio processing will be bypassed and any parameter changes will also be
bypassed.

Note:
setIsActive() isn't fully tested yet. It should be tested and possibly
improved, although it's a very simple design/concept.