Disable warnings and keep filter coefficients from updating every sample.
Disable warnings (in VS 2015) by either fixing the warning, or adding #pragmas
when needed (like for unreferenced parameters).

Also, comment out the calcFilter() function, which calculates the filter's coefficients,
in VAStateVariableFilter::processAudioBlock(). Disable this because it will calculate the
coefficients every sample and this is unneccesary work. This was used in processAudioBlock()
before because the smoother changed the coefficients and they needed to be updated as it
changed. The smoother was commented out though (it really isn't needed here).